Elaborate README concerning options for `run`
It seemed README was saying that if you specify `limit_count` it automatically
closes the circuit when `run` has been called the specified number of times.
But in reality it did not behave like that and kept the circuit open.

That is, the given example which was said to close the circuit after 1000 invocations
that looked like this

```ruby
CircuitSwitch.run(limit_count: 100) do
  # this proc is expected to be called not more than 100 times
end
```

does not actually close after 100 invocations.

In reality the desired behavior is fullfilled when and only when
`limit_count` is set along with `close_if_reach_limit: true`:

```ruby
CircuitSwitch.run(close_if_reach_limit: true, limit_count: 100) do
  # this proc is called not more than 100 times
end
```

Initially I thought this was a buggy behavior. But now I'm seeing it as a document's
flaw because the implementation clearly specifies `run_limit_count` is relevant
only when `close_if_reach_limit` is truthy:
https://github.com/makicamel/circuit_switch/blob/v0.2.2/lib/circuit_switch/core.rb#L21

Hence I'm presenting an elaboration on README so that it better represents the
expected behavior. For some stylistic reasons I have included changes that does not
directly associated with the concern described above, but I hope it displays
a better coherency in this vicinity.